requirements.txt
pytest==6.2.5
requests==2.27.1
cerberus==1.3.4
jsonschema==4.4.0
selenium==4.1.0
allure-pytest==2.9.45
pytest-xdist==2.5.0